[analyzer] Refine types returned by TypeSystemImpl methods.
In the following methods in `TypeSystemImpl`, the return type is
changed from `DartType` to `TypeImpl`:

- `greatestClosure`
- `greatestLowerBound`
- `leastClosure`
- `leastUpperBound`
- `makeNullable`
- `promoteToNonNull`

This required adding a few casts to `TypeSystemImpl`,
`LeastGreatestClosureHelper`, and `InterfaceTypeImpl`, which I believe
I will be able to remove in future CLs. It also allowed removing some
casts from `GreatestLowerBoundHelper`, `LeastUpperBoundHelper`, and
`InterfaceTypeImpl`.

This is part of a larger arc of work to change the analyzer's use of
the shared code so that the type parameters it supplies are not part
of the analyzer public API. See
https://github.com/dart-lang/sdk/issues/59763.

Change-Id: I24f27090bb2d07ae33be9c2c8d7908ef71a96929
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/404640
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>