[tests] Enum Shorthands - == and selector chain tests.
Similar format of tests to https://dart-review.googlesource.com/c/sdk/+/401980 and the other `==` and `!=` tests.

Making sure that equality with shorthand selector chains is allowed when the shorthand is on the RHS. Errors are produced when the shorthand is on the LHS, doesn't have the proper context, or has the context type of `Object`.

Bug: https://github.com/dart-lang/sdk/issues/57038
Change-Id: I905793b400a6f8b3c6bd068f70654dbbeec38557
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/403965
Reviewed-by: Erik Ernst <eernst@google.com>