[beta] Cherry pick DevTools 2.42.2.
Issue description: 1) The DevTools wasm build was crashing on exceptions because it was built with -O4 optimization. 2) DevTools extensions were broken because the extension asset path was being generated incorrectly.

What is the fix: 1) Build DevTools with -O2 optimization for the wasm
build and retain -O4 optimization for the dart2js build. 2) Use the
correct origin when generating DevTools asset paths.

Why cherry-pick: 1) We are crashing the wasm build for users trying this experiment and are unable to collect actionable data while monitoring the rollout. 2) The bug breaking DevTools extensions has not made it to stable yet, and it will without this cherry pick.

Risk: Low. 1) The wasm build of DevTools is still an opt-in experiment. 2) DevTools extensions have been tested and the fix is verified with this change.

Issue link(s): https://github.com/flutter/devtools/issues/8828

Cherry-pick: https://github.com/flutter/devtools/issues/8828
Change-Id: Ibf23e558d57cf26506c914efd953c75e44d7d9bd
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/406881
Reviewed-by: Elliott Brooks <elliottbrooks@google.com>
Commit-Queue: Kenzie Davisson <kenzieschmoll@google.com>
Reviewed-by: Ann Mossman <mossman@google.com>