[dart2wasm] Introduce unchecked entrypoints to instance members
If an instance member is a normal method or a setter, it has arguments
that may need to be type-checked. In some situations we know that we
don't have to perform them at all (e.g. if we know there's only
dispatches on `this`). In other situations we have guarantees on
individual call sites that we can skip the checks.

Up until now we have only taken advantage of the call site guarantees
when we decided to inline the target (then we avoided doing the type
checks).

In this CL we will make this also work if the target isn't inlined, but
called: Whenever a member has any parameters that we need to type check,
then we will generate checked & unchecked entrypoints. Both of them
do the optional parameter handling, but only the checked entrypoint will
perform the type checks, the unchecked entrypoint skips them. Both
unchecked & checked will call to a body function that has the body of
the member.

=> We will skip the type checks whether we inline the target or not.

The dart2wasm compiler currently represents targets it can call via
via `Reference`s: A member may be used in different ways: as a tear-off,
as a setter, type checker, etc.

=> We introduce now 3 more `Reference` kinds, namely checked, unchecked
and body.
=> The rest of the compiler is adjusted to also handle those new
`Reference` types.

All calls in the code generator that may target members that could have
checked and unchecked entrypoints now use
```
  Reference getFunctionEntry(Reference target, {required bool uncheckedEntry})
```

We maintain an invariant throughout the code base that a function

* **either** has only one "normal" entry if no arguments need type checks
* **or** has "unchecked" and "checked" entries (which both call a "body")

The dispatch table currently has only "normal" or "checked" entries in
it. So the "unchecked" entries are (if used) always called directly.

=> We only generate "unchecked" if there's direct unchecked calls.
=> We only generate "checked" if there's direct checked calls or
   dispatch table calls.
=> If only one entrypoint ends up being generated, binaryen can inline
   the body into the entrypoint function.
=> If both entrypoints are present, binaryen may often inline the
   unchecked one into call sites that then directly call the body.

In a future CL we may allow calling unchecked entries also via the
dispatch table.

Overall this approach leads to minimal changes to code size changes, but
brings -O2 performance closer to -O4.

Change-Id: Ic3082cc397335b969fd413f72652cc5af753adf7
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/406980
Reviewed-by: Ömer Ağacan <omersa@google.com>