[dart2js, ddc] Slightly quicker 'as' checks for primitve types
We have special methods in the rti library for checks like `as int`
and `as String?`. dart2js uses these when the type is known. This
change installs them on the `_as` property, so that we get a small
performance boost from calling a direct method rather than the `_as`
method calling the `_is` method.

Change-Id: Ia40a2fc7791f38590655a9b249fd9c36fb45369e
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/407604
Reviewed-by: Nicholas Shahan <nshahan@google.com>