[dart2wasm] Add support for dispatch table calls to unchecked entries
Now that we have support for generating checked & unchecked entrypoints,
we can make dispatch tables also target unchecked entry points.

This is beneficial especially in cases where there's dispatches on
`this` that require covariant checks but we don't know the target method (i.e. we cannot devirtualize it because the method we
dispatch to may beoverriden).

We keep the existing selectors that we have, but a selector will now
have

  * one row if none of the implementations of the selector need to
    perform type checks
    => `SelectorInfo` has a `SelectorTargets _normal`

  * two rows if any of the implementations of the selector need t
    perform
    a type check
    => `SelectorInfo` has a `SelectorTargets _checked`
    => `SelectorInfo` has a `SelectorTargets _unchecked`

Once an unchecked entrypoint is also used in the dispatch table (only
if there's any unchecked calls to that selector) then binaryen can no
longer optimize the signature of the function. It means we may have
perform e.g. downcasts / boxing in the unchecked entry where we
wouldn't do before (because we only had static calls to unchecked entry
before this PR).

So we're going to force-inline calls to unchecked entrypoints. This
avoids sometimes down casts and boxing. It also seems to actually
shrink the binary size.

Change-Id: I3ba4980c42886cc883fb610533f5fac9cce39b65
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/407740
Reviewed-by: Ömer Ağacan <omersa@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>