[ddc] Add web hot reload issue template.
This is a temporary template that we will use to capture feedback for stateful hot reload on the web.

Once hot reload is fully enabled and it is in a stable state, we will remove this template and any further issues discovered in the future will be filed via the normal Web issue filing template.

Change-Id: I462cb20898953f08335010f383574096315078c8
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/409020
Reviewed-by: Kevin Moore <kevmoo@google.com>
Commit-Queue: Nate Biggs <natebiggs@google.com>