[dart2wasm] Make json decoder construct [OneByteStrings] if possible
If the input to `utf8.decode()` has only a single non-ascii character we
currently make all strings [TwoByteString]s. Not only is this very
memory intensive, it's also going to make comparisons later on slower
(e.g. a json['foo'] will use [OneByteString] for `foo` but the actual
map will have a [TwoByteString] as `foo`).

In general it would be better to have an invariant that strings that can
be [OneByteString]s should be [OneByteString]s.

Change-Id: I740b3eeb0f3660f2bf8fdd094e12f78f318f9a5f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/409320
Reviewed-by: Ömer Ağacan <omersa@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>