[dart2wasm] Make json parser calculate string hash of keys while processing the json object key bytes
The json parser parses string literals in the json. If those string
literals are used as keys in json objects we'll later on construct a
hash map for the json object and have to calculate the hash of all the
string keys.

Since we process the bytes of the string key while scanning, we may as
well eagerly calculate the hash there while we already processed the
bytes, to avoid going over the bytes again when creating the hash map.

Change-Id: If998be31394594a64cc6c352edc6e9d805720356
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/409661
Reviewed-by: Ömer Ağacan <omersa@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>